feat(editProfile): renamed files to match Ifiok's
